requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Werkzeug

